Add a content metadata caching catalog.
The new build_content_catalog tool will build the intial catalog and and will update the catalog if changes are made to the content store outside of the included tools. All existing tools or library calls that the tools use have been updated to support the catalog. The real benefit of the catalog can be seen with either update_content or publish_content. Both tools are significantly faster now that the metadata is cached. However, there is still one major section in publish_content that needs to be optimized, but that is for another day.