Optimize the '_build_catalog' routine to use existing catalog data to speed up rebuilds.
The downside of the optimization is that the rebuild process will no longer fix certian types of cache corruption. To correct significant cache corruption it is recommend that you delete the current cache DB and start over. This is, in fact, effectively what the original version of '_build_catalog' did even though it was not visible to the user.

Also, some status messages where shifted around based on the optimization.